feat: Introduce the LinearStrGroupByKey/Mut types
